receiver: Uses uhd version of the usrp_rx_cfile
The capture.sh script was using usrp_rx_cfile which
is kind of outdated, it was replaced by the uhd_rc_cfile.
The .py suffix was removed since it is not used by gnuradio
anymore. DECIM is not used by uhd_rx_cfile instead it uses
the sample rate which is proper entered on this newest
version.

Signed-off-by: Felipe Zimmerle <felipe@zimmerle.org>